requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via streamlit
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via streamlit
google-ai-generativelanguage==0.6.4
    # via google-generativeai
google-api-core==2.28.1
    # via
    #   google-ai-generativelanguage
    #   google-api-python-client
    #   google-generativeai
google-api-python-client==2.187.0
    # via google-generativeai
google-auth==2.43.0
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-generativeai
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-generativeai==0.6.0
    # via calorie-estimator (pyproject.toml)
googleapis-common-protos==1.72.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.76.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.3
    # via google-api-core
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
idna==3.11
    # via requests
importlib-metadata==7.2.1
    # via streamlit
jinja2==3.1.6
    # via
    #   altair
    #   pydeck
jsonschema==4.25.1
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==2.13.0
    # via altair
numpy==1.26.4
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==23.2
    # via
    #   altair
    #   streamlit
pandas==2.3.3
    # via streamlit
pillow==10.2.0
    # via
    #   calorie-estimator (pyproject.toml)
    #   streamlit
proto-plus==1.26.1
    # via
    #   google-ai-generativelanguage
    #   google-api-core
protobuf==4.25.8
    # via
    #   google-ai-generativelanguage
    #   google-api-core
    #   google-generativeai
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
    #   streamlit
pyarrow==22.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydantic==2.12.5
    # via google-generativeai
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via rich
pyparsing==3.2.5
    # via httplib2
python-dateutil==2.9.0.post0
    # via
    #   pandas
    #   streamlit
python-dotenv==1.0.0
    # via calorie-estimator (pyproject.toml)
pytz==2025.2
    # via pandas
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   google-api-core
    #   streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
six==1.17.0
    # via python-dateutil
smmap==5.0.2
    # via gitdb
streamlit==1.31.1
    # via calorie-estimator (pyproject.toml)
tenacity==8.5.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.2
    # via streamlit
tqdm==4.67.1
    # via google-generativeai
typing-extensions==4.15.0
    # via
    #   altair
    #   google-generativeai
    #   grpcio
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
tzlocal==5.3.1
    # via streamlit
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
validators==0.35.0
    # via streamlit
zipp==3.23.0
    # via importlib-metadata
